% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitchapi_title_wordcloud.R
\name{top100_streams_title_wordcloud}
\alias{top100_streams_title_wordcloud}
\title{Word Cloud Title Search}
\usage{
top100_streams_title_wordcloud(x = 100, y = "en")
}
\arguments{
\item{x}{up to a maximum of 100 streams (100 default)}

\item{y}{language (english by defaults)}
}
\value{
A word cloud of most used terms in stream title
}
\description{
Collects top 100 stream titles and makes a word cloud of most popular words used
}
\examples{
top100_streams_title_wordcloud(100,"en")

}
\author{
Marcus Daffner
}
